pyproject.toml
[project]
name = "ml4sci"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    'black>=25.1.0',
    'h5py>=3.12',
    'matplotlib>3.5',
    'pillow>=11.1.0',
    'rich>=13.9.4',
    'scikit-learn>=1.6.1',
    'torch>=2.6.0',
    'torchinfo==1.8.0',
    'torchvision>=0.21',
    'torch-geometric>2.4'
]

